# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DISTRIMED Distribution Network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  superadmin: "superadmin",
  store: "store",
  customer: "customer",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  storeId?: number | null;
  active: boolean;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export interface Store {
  id: number;
  name: string;
  address: string;
  phone: string;
  email: string;
  lat: number;
  lng: number;
  active: boolean;
  description: string;
  imageUrl?: string | null;
  totalProducts: number;
  totalSales: number;
  createdAt: string;
}

export interface CreateStoreRequest {
  name: string;
  address: string;
  phone: string;
  email: string;
  lat: number;
  lng: number;
  description: string;
}

export interface UpdateStoreRequest {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  lat?: number;
  lng?: number;
  description?: string;
  active?: boolean;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl?: string | null;
  stock: number;
  storeId: number;
  storeName: string;
  avgRating: number;
  totalRatings: number;
  active: boolean;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  description: string;
  price: number;
  category: string;
  stock: number;
  storeId: number;
  imageUrl?: string;
}

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  stock?: number;
  imageUrl?: string;
  active?: boolean;
}

export interface Rating {
  id: number;
  productId: number;
  userId: number;
  stars: number;
  comment?: string | null;
  createdAt: string;
}

export interface CreateRatingRequest {
  productId: number;
  stars: number;
  comment?: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  storeName: string;
}

export interface StoreStub {
  id: number;
  name: string;
  totalSales: number;
  totalProducts: number;
}

export interface GlobalStats {
  totalStores: number;
  activeStores: number;
  totalProducts: number;
  totalUsers: number;
  totalSales: number;
  recentActivity: ActivityItem[];
  topStores: StoreStub[];
}

export interface SaleItem {
  id: number;
  productName: string;
  amount: number;
  quantity: number;
  timestamp: string;
}

export interface StoreStats {
  totalProducts: number;
  activeProducts: number;
  totalSales: number;
  avgRating: number;
  lowStockProducts: Product[];
  recentSales: SaleItem[];
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  superadmin: "superadmin",
  store: "store",
  customer: "customer",
} as const;

export interface CreateUserRequest {
  email: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  storeId?: number | null;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  superadmin: "superadmin",
  store: "store",
  customer: "customer",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  role?: UpdateUserRequestRole;
  active?: boolean;
  storeId?: number | null;
}

export type GetStoresParams = {
  active?: boolean;
};

export type GetProductsParams = {
  minPrice?: number;
  maxPrice?: number;
  category?: string;
  storeId?: number;
};