# lib/api-zod/src/generated/types/updateUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DISTRIMED Distribution Network API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserRequestRole } from "./updateUserRequestRole";

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  role?: UpdateUserRequestRole;
  active?: boolean;
  storeId?: number | null;
}